Updated use of port to simplify things while supporting more complicated stuff.
Modified do_install to save the installation port in .portid.
Modified commlib to default to the content of .portid.
Modified MainWindow.py to set the TTYPortRoot preference to
the port used by commlib, content of .portid initially.
Allow the user to specify a port in preferences to force use
of a different port in case more than one board is connected.
Changed the preferences filename from .animrc to .anim.prefs.